refactor: MenuGroupDao 제거
