+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<MenuGroup, Long> {
 }
