+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.menugroup.application;
 
 import kitchenpos.application.MenuGroupService;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,17 +26,17 @@ public class MenuGroupServiceTest {
     private MenuGroupService menuGroupService;
 
     @Mock
-    private MenuGroupDao menuGroupDao;
+    private MenuGroupRepository menuGroupRepository;
 
     @BeforeEach
     void setUp() {
-        menuGroupService = new MenuGroupService(menuGroupDao);
+        menuGroupService = new MenuGroupService(menuGroupRepository);
     }
 
     @DisplayName("메뉴 그룹을 생성한다.")
     @Test
     void create() {
-        when(menuGroupDao.save(any())).thenReturn(createMenuGroup01());
+        when(menuGroupRepository.save(any())).thenReturn(createMenuGroup01());
 
         // when
         MenuGroup created = menuGroupService.create(new MenuGroup(MENU_GROUP_NAME01));
@@ -49,7 +49,7 @@ public class MenuGroupServiceTest {
     @DisplayName("메뉴 그룹 목록을 조회한다.")
     @Test
     void list() {
-        when(menuGroupDao.findAll()).thenReturn(createMenuGroupList());
+        when(menuGroupRepository.findAll()).thenReturn(createMenuGroupList());
 
         // when
         List<MenuGroup> list = menuGroupService.list();
